requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi[standard]==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
django-environ==0.12.0
# Tika
tika-client==0.8.1
python-multipart==0.0.20
python-slugify==8.0.4
# Yara
yara-python==4.5.1
# Images
pillow==11.1.0
python-magic==0.4.27

